personaldiary-api/requirements.txt
flask
flask-smorest
pymongo
bson
flask_cors
boto3
python-dotenv
pathlib